+++ lib/slob_reader.dart
@@ -1,2 +1,5 @@
+library slob_reader;
+
 export 'src/models.dart';
 export 'src/reader.dart';
+export 'src/source.dart';